+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+Major changes are prefixed with `MAJOR`.
+
 ## [Unreleased]
 ### Added
 - New class `nochso\Omni\ArrayCollection` for creating collection classes.